Move SubtitleHelper to library
So that providers can use it when using the library dependency.
